refactor: split API methods into separate files and rename hooks for clarity
